:sparkles: Add libhal, util & mock to build & -fno-rtti
Having to add libhal, libhal-util and libhal-mock to each and every
library project that uses the cmake build utility is tedious boilerplate
that can be consolidated by this library. This will only affect
`libhal_test_and_make_library` as the others need to be free to be used
by `libhal`, `libhal-util`, and `libhal-mock` themselves.

Add -fexceptions & -fno-rtti flags to the library build flags to ensure
that unnecessary RTTI is not embedded into the library binaries.